+++ doc/releaseNotes.md
@@ -648,3 +648,7 @@ This file describes the main feature changes for each readout.exe released versi
 
 ## v2.26.3 - 22/10/2024
 - Minor release for osx compatibility.
+
+## v2.27.0 - 21/01/2024
+- Added symbolic links with relative path in configuration. See @LINK syntax.
+- Fixed o2-readout-config-generator for RHEL9 compatibility.